Made it so display button functions and added tab panel
I made it so when the display button is pressed it goes to the new panel panDisplay, I added different colors to the panels to show they are functioning when pressed as well. I removed the panelManInv panel and replaced it with tabManInv.  tabManInv is going to be a tabbed panel instead, under it I added a Add Inventory tab